Registrar alumno sin conexion
